% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGenRelatedness.R
\name{filterGenRelatedness}
\alias{filterGenRelatedness}
\title{Filter Feature Matrix Using Genetic Relatedness}
\usage{
filterGenRelatedness(featureMatrix, grm = NULL, thresholdP = 0.05)
}
\arguments{
\item{featureMatrix}{: a presence absence matrix where features are columns,
and samples are rows. The format should be a data frame and col/row names
should be preassigned.}

\item{grm}{: (default=NULL) if a GRM has been computed before for the
featureMatrix, it can be added here to by pass the computation.}

\item{thresholdP}{: (default=0.05) cut off p-value}
}
\value{
the GRM and a subsetted featureMatrix
}
\description{
This function computes the Genetic Relatedness Matrix (GRM) for a feature
matrix using minimal allele frequencies and extracts the first 10 principal
components of the GRM to regress against the features. The average of the
p-values is computed to filter out features with a high p-value indicating
correlation with population structure signal. This function can be used on
any feature data but theoretically it is most fit for Single Nucleotide
Polymorphism (SNP) data.
}
\examples{
filterGenRelatedness(featureMatrix = tinyFeatureMatrix)
}
