% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterProcAnalysis.R
\name{filterProcAnalysis}
\alias{filterProcAnalysis}
\title{Filter Feature Matrix Using Procrustes Distances}
\usage{
filterProcAnalysis(featureMatrix, phenotypes, threshold = 0.1)
}
\arguments{
\item{featureMatrix}{: a presence absence matrix where features are columns,
and samples are rows. The format should be a data frame and col/row names
should be preassigned.}

\item{phenotypes}{: a single column data frame of the phenotypes
assigned to each sample. Row names of this data frame should correspond to
that of the featureMatrix.}

\item{threshold}{: (default=0.1) Value between 0 and 1 that determines what
proportion of the features get filtered out. The features in the bottom
threshold*100% when sorted by Procrustes distances will be filtered out.}
}
\value{
a subset of the input featureMatrix
}
\description{
This function conducts Procrustes Analysis on feature matrices against
phenotype data to compute the Procrustes Distance of each feature. Features
with lower distances are more likely to be associated with population
structure signal, thus will be removed from the dataset. To further enhance
the efficacy of this function, dimensionality reduction can be done on the
featureMatrix and be used as input. The examples can be run if
TinySampleData.RData is loaded.
}
\examples{
filterProcAnalysis(featureMatrix = tinyFeatureMatrix,
phenotypes = phenotypes)

}
