% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBacPoSSR.R
\name{runBacPoSSR}
\alias{runBacPoSSR}
\title{Launch Shiny App for BacPoSSR}
\usage{
runBacPoSSR()
}
\value{
launches Shiny app, does not return any values.
}
\description{
This function launches the Shiny app for BacPoSSR. This app is intended for
those who are less accustomed to using R packages in code format. The app
should give you an interactive idea of what BacPoSSR can do.
The code for the app can be found in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

BacPoSSR::runBacPoSSR()
}

}
