% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultCompAnalysis.R
\name{plotMultCompAnalysis}
\alias{plotMultCompAnalysis}
\title{Plot Multiple Component Analysis (MCA)}
\usage{
plotMultCompAnalysis(
  featureMatrix,
  groups = NULL,
  saveTo = "./",
  title = "MCA"
)
}
\arguments{
\item{featureMatrix}{: a presence absence matrix where features are columns,
and samples are rows. The format should be a data frame and col/row names
should be preassigned.}

\item{groups}{: (default=NULL) a single column data frame of the groups
assigned to each sample. Row names of this data frame should correspond to
that of the featureMatrix.}

\item{saveTo}{: (default="./") path to directory where you want the MCA plot
to be saved.}

\item{title}{: (default="MCA") title of the plot.}
}
\value{
the MCA results
}
\description{
This function conducts Multiple Component Analysis (MCA) and plots the
resulting graph. This can be used to visually evaluate the efficacy of the
filtering methods provided in the BacPoSSR package or to visualize the
population structure signal of any given feature matrix. The examples can be
run if TinySampleData.RData is loaded.
}
\examples{
plotMultCompAnalysis(featureMatrix = tinyFeatureMatrix,
groups = tinyGroupsMatching)

}
\references{
Lê S, Josse J, Husson F (2008). “FactoMineR: A Package for Multivariate
Analysis.” Journal of Statistical Software, 25(1), 1–18.
doi:10.18637/jss.v025.i01.

Wickham H (2016). ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York. ISBN 978-3-319-24277-4,
https://ggplot2.tidyverse.org.
}
